makes send tx web API handler synced
